refactor(bitwise.ts, common.ts): extract common functions and constants to a separate file
